Move the validation package to utils as it's a generic library
